Controller fixes

+++ src/main/java/com/MasterBranch/repository/QueryRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.MasterBranch.model.Query;
 
-public interface QueryRepository  extends JpaRepository<Query, Long> {
+public interface QueryRepository  extends JpaRepository<Query, Integer> {
 	List<Query> findAll();
 }
